Yoseph discovered that FDR was not working properly.
there were 3 bugs in the code.
1) the parameters field in the output was not long enough, leading to "..."

2) the 'k' parameter of FDR correction was not properly initialized.

3) the scores were adjusted in the wrong direction, in the sense that 
the (k+1)-best seed was *guranteed* to have a score at least as good as the k-best seed.
the correct behavior is of course that the k-best seed should always (even after correction) have a better score than the (k+1)-best seed.
 
an appropriate unit test was added that checks for correct behavior.